refactor: refatora status
