+++ frontend/src/styles/App.css
@@ -72,23 +72,71 @@ select{
 
 .container-clients{
   display: flex;
-  justify-content: space-between; 
+  justify-content: space-between;
   align-items: center; 
   border: 0.5px solid #dbd8d8;
   padding: 5px;
   margin-bottom: 10px;
+  width: 100%;
 }
 
 
-.button-back{
+.aguardando{
+  margin: auto;
+  background-color: rgb(250, 155, 67);
+  border-radius: 50%;
+  width: 10px;
+  height: 10px;
+  margin-right: 15px;
+}
+
+.ativo{
+  margin: auto;
+  background-color: rgba(3, 170, 73, 0.795);
+  border-radius: 50%;
+  width: 10px;
+  height: 10px;
+  margin-right: 15px;
+}
+
+.inativo{
+  margin: auto;
+  background-color: rgba(161, 5, 18, 0.979);
+  border-radius: 50%;
+  width: 10px;
+  height: 10px;
+  margin-right: 15px;
+}
+
+.desativado{
+  margin: auto;
+  background-color: rgb(187, 184, 184);
+  border-radius: 50%;
+  width: 10px;
+  height: 10px;
+  margin-right: 15px;
+}
+
+.atualiza-status{
+  display: flex;
+}
+
+
+.button-edit{
   padding: 10px 30px;
   background-color: rgb(245, 241, 237);
   border: none;
   border-radius: 5px;
-  color: rgb(10, 10, 10);
+  color: rgb(250, 155, 67);
   cursor: pointer;
-  border: 0.5px solid #333333;
+  border: 0.5px solid rgb(250, 155, 67);
   margin-left: 35px;
+}
+
+.button-edit:hover{
+  background-color: rgb(250, 155, 67);
+  color: #dbd8d8;
+  border:rgb(250, 155, 67)
 }
 
 .form-buttons{
